big ol refactors again
