+++ serial-old/open_windows.go
@@ -15,10 +15,22 @@
 package serial
 
 import (
+	"fmt"
+	"io"
 	"os"
+	"sync"
 	"syscall"
 	"unsafe"
 )
+
+type serialPort struct {
+	f  *os.File
+	fd syscall.Handle
+	rl sync.Mutex
+	wl sync.Mutex
+	ro *syscall.Overlapped
+	wo *syscall.Overlapped
+}
 
 type structDCB struct {
 	DCBlength, BaudRate                            uint32
@@ -37,7 +49,7 @@ type structTimeouts struct {
 	WriteTotalTimeoutConstant   uint32
 }
 
-func openInternal(options OpenOptions) (*Port, error) {
+func openInternal(options OpenOptions) (io.ReadWriteCloser, error) {
 	if len(options.PortName) > 0 && options.PortName[0] != '\\' {
 		options.PortName = "\\\\.\\" + options.PortName
 	}
@@ -80,13 +92,51 @@ func openInternal(options OpenOptions) (*Port, error) {
 	if err != nil {
 		return nil, err
 	}
-	port := new(Port)
+	port := new(serialPort)
 	port.f = f
 	port.fd = h
 	port.ro = ro
 	port.wo = wo
 
 	return port, nil
+}
+
+func (p *serialPort) Close() error {
+	return p.f.Close()
+}
+
+func (p *serialPort) Write(buf []byte) (int, error) {
+	p.wl.Lock()
+	defer p.wl.Unlock()
+
+	if err := resetEvent(p.wo.HEvent); err != nil {
+		return 0, err
+	}
+	var n uint32
+	err := syscall.WriteFile(p.fd, buf, &n, p.wo)
+	if err != nil && err != syscall.ERROR_IO_PENDING {
+		return int(n), err
+	}
+	return getOverlappedResult(p.fd, p.wo)
+}
+
+func (p *serialPort) Read(buf []byte) (int, error) {
+	if p == nil || p.f == nil {
+		return 0, fmt.Errorf("Invalid port on read %v %v", p, p.f)
+	}
+
+	p.rl.Lock()
+	defer p.rl.Unlock()
+
+	if err := resetEvent(p.ro.HEvent); err != nil {
+		return 0, err
+	}
+	var done uint32
+	err := syscall.ReadFile(p.fd, buf, &done, p.ro)
+	if err != nil && err != syscall.ERROR_IO_PENDING {
+		return int(done), err
+	}
+	return getOverlappedResult(p.fd, p.ro)
 }
 
 var (
@@ -130,7 +180,7 @@ func setCommState(h syscall.Handle, options OpenOptions) error {
 	params.flags[0] = 0x01  // fBinary
 	params.flags[0] |= 0x10 // Assert DSR
 
-	if options.ParityMode != Parity_None {
+	if options.ParityMode != PARITY_NONE {
 		params.flags[0] |= 0x03 // fParity
 		params.Parity = byte(options.ParityMode)
 	}
@@ -240,6 +290,14 @@ func setCommMask(h syscall.Handle) error {
 	return nil
 }
 
+func resetEvent(h syscall.Handle) error {
+	r, _, err := syscall.Syscall(nResetEvent, 1, uintptr(h), 0, 0)
+	if r == 0 {
+		return err
+	}
+	return nil
+}
+
 func newOverlapped() (*syscall.Overlapped, error) {
 	var overlapped syscall.Overlapped
 	r, _, err := syscall.Syscall6(nCreateEvent, 4, 0, 1, 0, 0, 0, 0)
@@ -248,4 +306,17 @@ func newOverlapped() (*syscall.Overlapped, error) {
 	}
 	overlapped.HEvent = syscall.Handle(r)
 	return &overlapped, nil
+}
+
+func getOverlappedResult(h syscall.Handle, overlapped *syscall.Overlapped) (int, error) {
+	var n int
+	r, _, err := syscall.Syscall6(nGetOverlappedResult, 4,
+		uintptr(h),
+		uintptr(unsafe.Pointer(overlapped)),
+		uintptr(unsafe.Pointer(&n)), 1, 0, 0)
+	if r == 0 {
+		return n, err
+	}
+
+	return n, nil
 }
