+++ serial/interfaces.go
