+++ serial/open_linux.go
@@ -2,32 +2,14 @@ package serial
 
 import (
 	"errors"
-	"io"
 	"os"
-	"syscall"
 	"unsafe"
 
 	"golang.org/x/sys/unix"
 )
 
-//
-// Grab the constants with the following little program, to avoid using cgo:
-//
-// #include <stdio.h>
-// #include <stdlib.h>
-// #include <linux/termios.h>
-//
-// int main(int argc, const char **argv) {
-//   printf("TCSETS2 = 0x%08X\n", TCSETS2);
-//   printf("BOTHER  = 0x%08X\n", BOTHER);
-//   printf("NCCS    = %d\n",     NCCS);
-//   return 0;
-// }
-//
 const (
-	kTCSETS2 = 0x402C542B
-	kBOTHER  = 0x1000
-	kNCCS    = 19
+	NCCS = 19
 )
 
 //
@@ -38,31 +20,14 @@ type cc_t byte
 type speed_t uint32
 type tcflag_t uint32
 type termios2 struct {
-	c_iflag  tcflag_t    // input mode flags
-	c_oflag  tcflag_t    // output mode flags
-	c_cflag  tcflag_t    // control mode flags
-	c_lflag  tcflag_t    // local mode flags
-	c_line   cc_t        // line discipline
-	c_cc     [kNCCS]cc_t // control characters
-	c_ispeed speed_t     // input speed
-	c_ospeed speed_t     // output speed
-}
-
-// Constants for RS485 operation
-
-const (
-	sER_RS485_ENABLED        = (1 << 0)
-	sER_RS485_RTS_ON_SEND    = (1 << 1)
-	sER_RS485_RTS_AFTER_SEND = (1 << 2)
-	sER_RS485_RX_DURING_TX   = (1 << 4)
-	tIOCSRS485               = 0x542F
-)
-
-type serial_rs485 struct {
-	flags                 uint32
-	delay_rts_before_send uint32
-	delay_rts_after_send  uint32
-	padding               [5]uint32
+	c_iflag  tcflag_t   // input mode flags
+	c_oflag  tcflag_t   // output mode flags
+	c_cflag  tcflag_t   // control mode flags
+	c_lflag  tcflag_t   // local mode flags
+	c_line   cc_t       // line discipline
+	c_cc     [NCCS]cc_t // control characters
+	c_ispeed speed_t    // input speed
+	c_ospeed speed_t    // output speed
 }
 
 //
@@ -85,12 +50,12 @@ func makeTermios2(options OpenOptions) (*termios2, error) {
 		return nil, errors.New("invalid value for InterCharacterTimeout")
 	}
 
-	ccOpts := [kNCCS]cc_t{}
-	ccOpts[syscall.VTIME] = cc_t(vtime / 100)
-	ccOpts[syscall.VMIN] = cc_t(vmin)
+	ccOpts := [NCCS]cc_t{}
+	ccOpts[unix.VTIME] = cc_t(vtime / 100)
+	ccOpts[unix.VMIN] = cc_t(vmin)
 
 	t2 := &termios2{
-		c_cflag:  syscall.CLOCAL | syscall.CREAD | kBOTHER,
+		c_cflag:  unix.CLOCAL | unix.CREAD | unix.BOTHER,
 		c_ispeed: speed_t(options.BaudRate),
 		c_ospeed: speed_t(options.BaudRate),
 		c_cc:     ccOpts,
@@ -99,20 +64,20 @@ func makeTermios2(options OpenOptions) (*termios2, error) {
 	switch options.StopBits {
 	case 1:
 	case 2:
-		t2.c_cflag |= syscall.CSTOPB
+		t2.c_cflag |= unix.CSTOPB
 
 	default:
 		return nil, errors.New("invalid setting for StopBits")
 	}
 
 	switch options.ParityMode {
-	case PARITY_NONE:
-	case PARITY_ODD:
-		t2.c_cflag |= syscall.PARENB
-		t2.c_cflag |= syscall.PARODD
+	case Parity_None:
+	case Parity_Odd:
+		t2.c_cflag |= unix.PARENB
+		t2.c_cflag |= unix.PARODD
 
-	case PARITY_EVEN:
-		t2.c_cflag |= syscall.PARENB
+	case Parity_Even:
+		t2.c_cflag |= unix.PARENB
 
 	default:
 		return nil, errors.New("invalid setting for ParityMode")
@@ -120,37 +85,32 @@ func makeTermios2(options OpenOptions) (*termios2, error) {
 
 	switch options.DataBits {
 	case 5:
-		t2.c_cflag |= syscall.CS5
+		t2.c_cflag |= unix.CS5
 	case 6:
-		t2.c_cflag |= syscall.CS6
+		t2.c_cflag |= unix.CS6
 	case 7:
-		t2.c_cflag |= syscall.CS7
+		t2.c_cflag |= unix.CS7
 	case 8:
-		t2.c_cflag |= syscall.CS8
+		t2.c_cflag |= unix.CS8
 	default:
 		return nil, errors.New("invalid setting for DataBits")
-	}
-
-	if options.RTSCTSFlowControl {
-		t2.c_cflag |= unix.CRTSCTS
 	}
 
 	return t2, nil
 }
 
-func openInternal(options OpenOptions) (io.ReadWriteCloser, error) {
-
+func openInternal(options OpenOptions) (*Port, error) {
 	file, openErr :=
 		os.OpenFile(
 			options.PortName,
-			syscall.O_RDWR|syscall.O_NOCTTY|syscall.O_NONBLOCK,
+			unix.O_RDWR|unix.O_NOCTTY|unix.O_NONBLOCK,
 			0600)
 	if openErr != nil {
 		return nil, openErr
 	}
 
 	// Clear the non-blocking flag set above.
-	nonblockErr := syscall.SetNonblock(int(file.Fd()), false)
+	nonblockErr := unix.SetNonblock(int(file.Fd()), false)
 	if nonblockErr != nil {
 		return nil, nonblockErr
 	}
@@ -160,10 +120,10 @@ func openInternal(options OpenOptions) (io.ReadWriteCloser, error) {
 		return nil, optErr
 	}
 
-	r, _, errno := syscall.Syscall(
-		syscall.SYS_IOCTL,
+	r, _, errno := unix.Syscall(
+		unix.SYS_IOCTL,
 		uintptr(file.Fd()),
-		uintptr(kTCSETS2),
+		uintptr(unix.TCSETS2),
 		uintptr(unsafe.Pointer(t2)))
 
 	if errno != 0 {
@@ -174,36 +134,5 @@ func openInternal(options OpenOptions) (io.ReadWriteCloser, error) {
 		return nil, errors.New("unknown error from SYS_IOCTL")
 	}
 
-	if options.Rs485Enable {
-		rs485 := serial_rs485{
-			sER_RS485_ENABLED,
-			uint32(options.Rs485DelayRtsBeforeSend),
-			uint32(options.Rs485DelayRtsAfterSend),
-			[5]uint32{0, 0, 0, 0, 0},
-		}
-
-		if options.Rs485RtsHighDuringSend {
-			rs485.flags |= sER_RS485_RTS_ON_SEND
-		}
-
-		if options.Rs485RtsHighAfterSend {
-			rs485.flags |= sER_RS485_RTS_AFTER_SEND
-		}
-
-		r, _, errno := syscall.Syscall(
-			syscall.SYS_IOCTL,
-			uintptr(file.Fd()),
-			uintptr(tIOCSRS485),
-			uintptr(unsafe.Pointer(&rs485)))
-
-		if errno != 0 {
-			return nil, os.NewSyscallError("SYS_IOCTL (RS485)", errno)
-		}
-
-		if r != 0 {
-			return nil, errors.New("Unknown error from SYS_IOCTL (RS485)")
-		}
-	}
-
-	return file, nil
+	return NewPort(file), nil
 }
